requirements.txt
discord.py==1.3.4
discord==1.7.3
python-dotenv==0.19.1
